--- src/refactoring/entities/cart/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-cart";
-export * from "../../../contexts/cart-context";
