--- src/refactoring/entities/coupon/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-coupon";
-export * from "../../../contexts/coupon-context";
