--- src/refactoring/entities/product/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-product";
-export * from "./product-context";
